\name{ESkNN-package}
\alias{ESkNN-package}
\alias{ESkNN}
\docType{package}
\title{ Ensemble of Subset of K-Nearest Neighbours Classifiers for Classification and Class Membership Probability Estimation
}
\description{Functions for building an ensemble of optimal k-nearest neighbours (kNN) models for classification and class membership probability estimation are provided.  To address the issue of non-informative features in the data. A set of base kNN models is generated and a subset of models is selected for the ensemble based on the individual and combined performance of these models. Out-of-bag data and an independent training data set is used for the performance assessment of models individually and collectively. Class labels and class membership probability estimates are returned by the prediction functions. Other measures such as confusion matrix, classification error rate, and brier scores etc, are also returned by the functions.
}
\details{
\tabular{ll}{
Package: \tab ESKNN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-13\cr
License: \tab GPL (>= 2)\cr
}

}
\author{Asma Gul, Aris Perperoglou, Zardad Khan, Osama Mahmoud, Miftahuddin, Werner Adler, and Berthold Lausen
Maintainer: Asma Gul <agul@essex.ac.uk>
}
\references{Gul, A., Perperoglou, A., Khan, Z., Mahmoud, O., Miftahuddin, M., Adler, W. and Lausen, B.(2014),\emph{Ensemble of subset of k-nearest neighbours classifiers}, Journal name to appear.}
\keyword{ package }
